+++ src/game/inventry.c
@@ -697,6 +697,17 @@ int32_t Display_Inventory(int inv_mode)
                 // page 3: exit game
                 return GF_EXIT_GAME;
             }
+        } else if (g_InvMode == INV_DEATH_MODE) {
+            if (g_InvExtraData[0] == 0) {
+                // page 1: load game
+                return GF_START_SAVED_GAME | g_InvExtraData[1];
+            } else if (g_InvExtraData[0] == 1) {
+                // page 2: restart level
+                return GF_START_GAME | g_CurrentLevel;
+            } else {
+                // page 3: exit game
+                return GF_EXIT_TO_TITLE;
+            }
         } else {
             if (g_InvExtraData[0] == 0) {
                 // page 1: load game